maj sécurité route delete

+++ back-end/controllers/sauce.js
@@ -46,11 +46,16 @@ exports.deleteSauce = (req, res, next) => {
         _id: req.params.id
     }).then(sauce => {
         const filename = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.deleteOne({ _id:req.params.id })
-            .then(() => res.status(200).json({message: 'Objet supprimé !'}))
-            .catch(error => res.status(400).json({error }));
-        });
+        if(sauce.userId === req.auth.userId){
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.deleteOne({ _id:req.params.id })
+                .then(() => res.status(200).json({message: 'Objet supprimé !'}))
+                .catch(error => res.status(400).json({error }));
+            });
+        }else{
+            throw "userId différent de userId objet à supprimer"
+        }
+
     })
     .catch(error => res.status(500).json({ error}));
 };
